Refactor REGON API provider into modular components
- Restructured the REGON API provider by breaking it down into multiple components including `RegonApiClient`, `RegonDataMapper`, `RegonRateLimiter`, and `RegonSessionManager`.
- Introduced a new `RegonSoapClient` for handling SOAP requests.
- Updated the main `RegonProvider` class to utilize these components for improved maintainability and clarity.
- Added an `__init__.py` file to the `regon` module for better organization and accessibility of components.
